+++ src/rpc/rpc_port_load_request.cpp
@@ -1,4 +1,5 @@
 #include "rpc_port_load_request.h"
+#include "rpc_port_driver_list.h"
 
 namespace
 {
