+++ src/rpc/rpc_port_scan_serial_client_task.cpp
@@ -1,7 +1,7 @@
 #include "rpc_port_scan_serial_client_task.h"
 #include "modbus_base.h"
 #include "modbus_common.h"
-#include "rpc_port_handler.h"
+#include "rpc_port_driver_list.h"
 #include "rpc_port_scan_handler.h"
 #include "serial_exc.h"
 #include "serial_port.h"
