+++ src/rpc/rpc_port_setup_handler.cpp
@@ -1,5 +1,5 @@
 #include "rpc_port_setup_handler.h"
-#include "rpc_port_handler.h"
+#include "rpc_port_driver_list.h"
 #include "rpc_port_setup_serial_client_task.h"
 #include "serial_port.h"
 #include "tcp_port.h"
